significant refactoring
